extract remaining ssv-network helpers
